Split SSL key-share tests
